separate adapters to own files
